+++ src/Console/tests/Fixtures/Status.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Spiral\Tests\Console\Fixtures;
+
+enum Status: int
+{
+    case Active = 1;
+    case Inactive = 0;
+}